requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
plotly==5.18.0
mediapipe==0.10.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
flask-login==0.6.3
PyJWT==2.8.0
passlib==1.7.4
psutil==5.9.8
flask-sqlalchemy==3.1.1
face-recognition==1.3.0
torch==2.1.2
torchvision==0.16.2
